separate classes, unified interface
