+++ barretenberg/cpp/src/barretenberg/circuit_checker/ultra_circuit_checker.cpp
@@ -1,21 +1,21 @@
-#include "circuit_checker.hpp"
+#include "ultra_circuit_checker.hpp"
 #include "barretenberg/flavor/goblin_ultra.hpp"
 #include <barretenberg/plonk/proof_system/constants.hpp>
 #include <unordered_set>
 
 namespace bb {
 
-template <> auto CircuitChecker::init_empty_values<UltraCircuitBuilder_<UltraArith<bb::fr>>>()
+template <> auto UltraCircuitChecker::init_empty_values<UltraCircuitBuilder_<UltraArith<bb::fr>>>()
 {
     return UltraFlavor::AllValues{};
 }
 
-template <> auto CircuitChecker::init_empty_values<GoblinUltraCircuitBuilder_<bb::fr>>()
+template <> auto UltraCircuitChecker::init_empty_values<GoblinUltraCircuitBuilder_<bb::fr>>()
 {
     return GoblinUltraFlavor::AllValues{};
 }
 
-template <typename Builder> bool CircuitChecker::check(const Builder& builder_in)
+template <typename Builder> bool UltraCircuitChecker::check(const Builder& builder_in)
 {
     // Create a copy of the input circuit and finalize it
     Builder builder{ builder_in };
@@ -58,11 +58,11 @@ template <typename Builder> bool CircuitChecker::check(const Builder& builder_in
 };
 
 template <typename Builder>
-bool CircuitChecker::check_block(Builder& builder,
-                                 auto& block,
-                                 TagCheckData& tag_data,
-                                 MemoryCheckData& memory_data,
-                                 LookupHashTable& lookup_hash_table)
+bool UltraCircuitChecker::check_block(Builder& builder,
+                                      auto& block,
+                                      TagCheckData& tag_data,
+                                      MemoryCheckData& memory_data,
+                                      LookupHashTable& lookup_hash_table)
 {
     // Initialize empty AllValues of the correct Flavor based on Builder type; for input to Relation::accumulate
     auto values = init_empty_values<Builder>();
@@ -121,7 +121,7 @@ bool CircuitChecker::check_block(Builder& builder,
     return result;
 };
 
-template <typename Relation> bool CircuitChecker::check_relation(auto& values, auto& params)
+template <typename Relation> bool UltraCircuitChecker::check_relation(auto& values, auto& params)
 {
     // Define zero initialized array to store the evaluation of each sub-relation
     using SubrelationEvaluations = typename Relation::SumcheckArrayOfValuesOverSubrelations;
@@ -142,7 +142,7 @@ template <typename Relation> bool CircuitChecker::check_relation(auto& values, a
     return true;
 }
 
-bool CircuitChecker::check_lookup(auto& values, auto& lookup_hash_table)
+bool UltraCircuitChecker::check_lookup(auto& values, auto& lookup_hash_table)
 {
     // If this is a lookup gate, check the inputs are in the hash table containing all table entries
     if (!values.q_lookup.is_zero()) {
@@ -154,13 +154,13 @@ bool CircuitChecker::check_lookup(auto& values, auto& lookup_hash_table)
     return true;
 };
 
-bool CircuitChecker::check_tag_data(const TagCheckData& tag_data)
+bool UltraCircuitChecker::check_tag_data(const TagCheckData& tag_data)
 {
     return tag_data.left_product == tag_data.right_product;
 };
 
 template <typename Builder>
-void CircuitChecker::populate_values(
+void UltraCircuitChecker::populate_values(
     Builder& builder, auto& block, auto& values, TagCheckData& tag_data, MemoryCheckData& memory_data, size_t idx)
 {
     // Function to quickly update tag products and encountered variable set by index and value
@@ -245,11 +245,8 @@ void CircuitChecker::populate_values(
 }
 
 // Template method instantiations for each check method
-// template bool CircuitChecker::check<bb::fr>(const StandardCircuitBuilder_<bb::fr>& builder);
-// template bool CircuitChecker::check<bb::fq>(const StandardCircuitBuilder_<bb::fq>& builder);
-template bool CircuitChecker::check<UltraCircuitBuilder_<UltraArith<bb::fr>>>(
+template bool UltraCircuitChecker::check<UltraCircuitBuilder_<UltraArith<bb::fr>>>(
     const UltraCircuitBuilder_<UltraArith<bb::fr>>& builder_in);
-template bool CircuitChecker::check<GoblinUltraCircuitBuilder_<bb::fr>>(
+template bool UltraCircuitChecker::check<GoblinUltraCircuitBuilder_<bb::fr>>(
     const GoblinUltraCircuitBuilder_<bb::fr>& builder_in);
-
 } // namespace bb
